+++ cedar-db/src/expr_to_query.rs
@@ -1,9 +1,9 @@
-use cedar_policy::{EntityTypeName, Schema, RandomRequestEnv, ResidualResponse, Effect};
-use cedar_policy_core::ast::{Literal, UnaryOp, BinaryOp, Expr, ExprBuilder};
-use cedar_policy_validator::{typecheck::Typechecker, ValidationMode};
-use sea_query::{SimpleExpr, IntoColumnRef, BinOper, extension::postgres::{PgBinOper, PgExpr}, Alias, IntoIden, Query, Keyword, PgFunc, SelectStatement, IntoTableRef};
+use cedar_policy::EntityTypeName;
+use cedar_policy_core::ast::{Literal, UnaryOp, BinaryOp};
+use sea_query::{SimpleExpr, IntoColumnRef, BinOper, extension::postgres::{PgBinOper, PgExpr}, Alias, IntoIden, Query, Keyword, PgFunc};
 
-use crate::query_expr::{QueryExprError, QueryExpr, UnknownType, CastableType, ExprWithBindings, IdGen, AttrOrId};
+
+use crate::query_expr::{QueryExprError, QueryExpr, UnknownType, CastableType, AttrOrId};
 
 type Result<T> = std::result::Result<T, QueryExprError>;
 
@@ -231,258 +231,5 @@ impl QueryExpr {
             QueryExpr::Set(_) => unimplemented!("TODO: implement Set"),
             QueryExpr::Record { .. } => unimplemented!("TODO: implement Record"),
         }
-    }
-}
-
-impl ExprWithBindings {
-    pub fn to_sql_query<T: IntoTableRef, U: IntoIden>(&self, ein: impl InConfig, table_names: impl Fn(&EntityTypeName) -> (T, U)) -> Result<SelectStatement> {
-        let mut query = Query::select();
-        query.and_where(self.expr.to_sql_query(&ein)?);
-        for (bv, e) in self.bindings.iter() {
-            let (tbl_ref, id_name) = table_names(&bv.ty);
-            let id_name = id_name.into_iden();
-            query.join_as(sea_query::JoinType::InnerJoin,
-                tbl_ref, Alias::new(bv.name.clone()),
-                e.to_sql_query(&ein)?.eq((Alias::new(bv.name.clone()), id_name).into_column_ref()));
-        }
-        Ok(query)
-    }
-
-    pub fn add_froms<T: IntoTableRef, U: IntoIden>(&self, query: &mut SelectStatement, table_names: impl Fn(&EntityTypeName) -> (T, U)) {
-        let unks = self.get_free_vars();
-        for (unk, tp) in unks {
-            if let Some(ty) = tp {
-                let (tbl_ref, _) = table_names(ty);
-                query.from_as(tbl_ref, Alias::new(unk.name.as_str()));  // todo: cross join any further tables
-            }
-        }
-    }
-
-    pub fn add_ids<T: IntoTableRef, U: IntoIden>(&self, query: &mut SelectStatement, table_names: impl Fn(&EntityTypeName) -> (T, U)) {
-        let unks = self.get_free_vars();
-        for (unk, tp) in unks {
-            if let Some(ty) = tp {
-                let (_, id_name) = table_names(ty);
-                query.column((Alias::new(unk.name.as_str()), id_name.into_iden()));
-            }
-        }
-    }
-}
-
-/// Does the translation from Cedar to SQL
-pub fn translate_expr<T: IntoTableRef, U: IntoIden>(expr: &Expr, schema: &Schema, ein: impl InConfig, table_names: impl Fn(&EntityTypeName) -> (T, U)) -> Result<SelectStatement> {
-    let typechecker = Typechecker::new(&schema.0, ValidationMode::Strict);
-    let req_env = RandomRequestEnv::new();
-    let typed_expr = typechecker.typecheck_expr_strict(&(&req_env).into(), expr, cedar_policy_validator::types::Type::primitive_boolean(), &mut Vec::new())
-        .ok_or(QueryExprError::TypecheckError)?;
-
-    let query_expr: QueryExpr = (&typed_expr).try_into()?;
-
-    let mut query_with_bindings: ExprWithBindings = query_expr.into();
-    query_with_bindings.reduce_attrs(&mut IdGen::new());
-
-    query_with_bindings.to_sql_query(ein, table_names)
-}
-
-
-pub fn translate_response<T: IntoTableRef, U: IntoIden>(resp: &ResidualResponse, schema: &Schema, ein: impl InConfig, table_names: impl Fn(&EntityTypeName) -> (T, U)) -> Result<SelectStatement> {
-    let (permits, forbids): (Vec<_>, Vec<_>) =
-        resp.residuals().policies()
-            .partition(|p| p.effect() == Effect::Permit);
-    let expr: Expr = ExprBuilder::new().and(
-        ExprBuilder::new().any(permits.into_iter().map(|p| p.non_head_constraints().clone())),
-        ExprBuilder::new().not(ExprBuilder::new().any(forbids.into_iter().map(|p| p.non_head_constraints().clone()))));
-    let query = translate_expr(&expr, schema, ein, table_names)?;
-    Ok(query)
-}
-
-#[cfg(test)]
-mod test {
-    use cedar_policy::Schema;
-    use cedar_policy_core::{evaluator::RestrictedEvaluator, extensions::Extensions, ast};
-    use sea_query::{Alias, PostgresQueryBuilder, Asterisk};
-
-    use super::{translate_expr, InByTable};
-
-    /// Translation function for the purposes of testing; fills in lots of boilerplate
-    pub fn translate_expr_test(expr: ast::Expr, schema: &Schema) -> String {
-        let ext = Extensions::all_available();
-        let eval = RestrictedEvaluator::new(&ext);
-        let expr = eval.partial_interpret_unrestricted(&expr, &["unknown".parse().unwrap()].into()).unwrap();
-
-        let mut query = translate_expr(&expr, schema, &InByTable(|t1, t2| {
-            let t1_str = t1.to_string();
-            let t2_str = t2.to_string();
-            let in_table = format!("{}_in_{}", t1_str, t2_str);
-            Ok((Alias::new(in_table), Alias::new(t1_str), Alias::new(t2_str)))
-        }), |tp| (Alias::new(tp.basename()), Alias::new("uid"))).unwrap();
-
-        query
-            .column(Asterisk)
-            .from(Alias::new("resource"))
-            .to_string(PostgresQueryBuilder)
-    }
-
-
-    fn get_schema() -> Schema {
-        r#"
-        {
-            "": {
-                "entityTypes": {
-                    "Users": {
-                        "shape": {
-                            "type": "Record",
-                            "attributes": {
-                                "level": {
-                                    "type": "Long"
-                                },
-                                "info": {
-                                    "type": "Record",
-                                    "attributes": {
-                                        "name": {
-                                            "type": "String"
-                                        },
-                                        "age": {
-                                            "type": "Long"
-                                        }
-                                    }
-                                },
-                                "allowedPhotos": {
-                                    "type": "Set",
-                                    "element": {
-                                        "type": "Entity",
-                                        "name": "Photos"
-                                    }
-                                }
-                            }
-                        }
-                    },
-                    "Photos": {
-                        "shape": {
-                            "type": "Record",
-                            "attributes": {
-                                "owner": {
-                                    "type": "Entity",
-                                    "name": "Users"
-                                },
-                                "nextPhoto": {
-                                    "type": "Entity",
-                                    "name": "Photos"
-                                }
-                            }
-                        },
-                        "memberOfTypes": ["Photos"]
-                    }
-                },
-                "actions": {
-                    "view": {
-                        "appliesTo": {
-                            "principalTypes": ["Users"],
-                            "resourceTypes": ["Photos"]
-                        }
-                    }
-                }
-            }
-        }
-        "#.parse().unwrap()
-    }
-
-    #[test]
-    fn test_basic() {
-        let result = translate_expr_test(
-            r#"unknown("resource: Photos").owner == Users::"0""#.parse().unwrap(),
-            &get_schema(),
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE "resource"."owner" = '0'"#);
-    }
-
-    #[test]
-    fn test_entity_deref_id() {
-        let result = translate_expr_test(
-            r#"unknown("resource: Photos") == Photos::"0""#.parse().unwrap(),
-            &get_schema(),
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE "resource"."uid" = '0'"#);
-    }
-
-    #[test]
-    fn test_if() {
-        let result = translate_expr_test(
-            r#"(if unknown("resource: Photos").owner == Users::"0" then Photos::"beach" else Photos::"carnival") == unknown("resource: Photos").nextPhoto"#.parse().unwrap(),
-            &get_schema(),
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE (CASE WHEN ("resource"."owner" = '0') THEN 'beach' ELSE 'carnival' END) = "resource"."nextPhoto""#);
-    }
-
-    #[test]
-    fn test_nested_getattr() {
-        let result = translate_expr_test(
-            r#"5 <= unknown("resource: Photos").owner.level"#.parse().unwrap(),
-            &get_schema(),
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" INNER JOIN "Users" AS "v__entity_attr_0" ON "resource"."owner" = "v__entity_attr_0"."uid" WHERE 5 <= "v__entity_attr_0"."level""#);
-    }
-
-    #[test]
-    fn test_deeply_nested_getattr() {
-        let result = translate_expr_test(
-            r#"unknown("resource: Photos").nextPhoto.nextPhoto.nextPhoto.owner.level == 3"#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" INNER JOIN "Photos" AS "v__entity_attr_0" ON "resource"."nextPhoto" = "v__entity_attr_0"."uid" INNER JOIN "Photos" AS "v__entity_attr_1" ON "v__entity_attr_0"."nextPhoto" = "v__entity_attr_1"."uid" INNER JOIN "Photos" AS "v__entity_attr_2" ON "v__entity_attr_1"."nextPhoto" = "v__entity_attr_2"."uid" INNER JOIN "Users" AS "v__entity_attr_3" ON "v__entity_attr_2"."owner" = "v__entity_attr_3"."uid" WHERE "v__entity_attr_3"."level" = 3"#);
-    }
-
-    #[test]
-    fn test_double_getattr() {
-        let result = translate_expr_test(
-            r#"5 <= unknown("resource: Photos").owner.level && unknown("resource: Photos").owner.level <= 10"#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" INNER JOIN "Users" AS "v__entity_attr_0" ON "resource"."owner" = "v__entity_attr_0"."uid" WHERE (5 <= "v__entity_attr_0"."level") AND ("v__entity_attr_0"."level" <= 10)"#);
-    }
-
-    #[test]
-    fn test_if_getattr() {
-        let result = translate_expr_test(
-            r#"(if unknown("resource: Photos") == Photos::"0" || unknown("resource: Photos") == Photos::"1" then unknown("resource: Photos") else unknown("resource: Photos").nextPhoto).owner == Users::"2""#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" INNER JOIN "Photos" AS "v__entity_attr_0" ON (CASE WHEN (("resource"."uid" = '0') OR ("resource"."uid" = '1')) THEN "resource"."uid" ELSE "resource"."nextPhoto" END) = "v__entity_attr_0"."uid" WHERE "v__entity_attr_0"."owner" = '2'"#);
-    }
-
-    #[test]
-    fn test_in() {
-        let result: String = translate_expr_test(
-            r#"unknown("resource: Photos") in Users::"0""#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE '0' IN (SELECT "Photos_in_Users"."Users" FROM "Photos_in_Users" WHERE "Photos_in_Users"."Photos" = "resource"."uid")"#);
-    }
-
-    #[test]
-    fn test_in2() {
-        let result: String = translate_expr_test(
-            r#"Users::"0" in unknown("resource: Photos")"#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE "resource"."uid" IN (SELECT "Users_in_Photos"."Photos" FROM "Users_in_Photos" WHERE "Users_in_Photos"."Users" = '0')"#);
-    }
-
-    #[test]
-    fn test_inset() {
-        let result: String = translate_expr_test(
-            r#"Photos::"0" in unknown("resource: Users").allowedPhotos"#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" WHERE ('0' = ANY("resource"."allowedPhotos")) OR ('0' IN (SELECT "Photos_in_Photos"."Photos" FROM "Photos_in_Photos" WHERE "Photos_in_Photos"."Photos" = ANY("resource"."allowedPhotos")))"#)
-    }
-
-    #[test]
-    fn test_record_attr() {
-        let result: String = translate_expr_test(
-            r#"unknown("resource: Photos").owner.info.name == "Bob""#.parse().unwrap(),
-            &get_schema()
-        );
-        assert_eq!(result, r#"SELECT * FROM "resource" INNER JOIN "Users" AS "v__entity_attr_0" ON "resource"."owner" = "v__entity_attr_0"."uid" WHERE "v__entity_attr_0"."info" ->> 'name' = 'Bob'"#)
     }
 }
